Move tcp logic to module tcp
